Fix failing tests after upgrading to Faraday version 2.0.0
Some client tests were failing with this error:
NoMethodError: undefined method `key?' for "{}":String (lib/cent/http.rb:42)

The failure is due to the difference between Faraday middleware's JSON `#parse_response?` implementation,
and Faraday's version 2.0.0 implementation of the now built-in `#parse_response?`.

The middleware parses the body if the body starts with a bracket.
https://github.com/lostisland/faraday_middleware/blob/3f6b113f6a7f71ef05c35a03514bbc08f7993dcf/lib/faraday_middleware/response/parse_json.rb#L35

While, Faraday parses the body if the response Content-Type header matches the configured value. (Default: /\bjson$/)
https://github.com/lostisland/faraday/blob/4816043101ea57f806d0946ed29f252e06d16909/lib/faraday/response/json.rb#L38

This commit fixes the error by adding a "Content-Type: application/json" header to the stubbed response.